Moved more code away from main symbolicator module
